+++ ops/activation/hardswish.cpp
@@ -20,8 +20,8 @@ void kernel(float input[], float output[])
     #pragma clang loop unroll_count(1) vectorize(disable)//vectorize_width(4)
     for (int i = 0; i < NTAPS; i++) {
         float x = input[i];
-        if (x <= -3.0) output[i] = 0.0;
-        else if (x >= 3.0) output[i] = x;
-        else output[i] = x * (x + 3.0) / 6.0;
+        if (x <= (float)(-3.0)) output[i] = (float)(0.0);
+        else if (x >= (float)(3.0)) output[i] = x;
+        else output[i] = x * (x + (float)(3.0)) / (float)(6.0);
     }
 }
