+++ ops/activation/selu.cpp
@@ -22,6 +22,6 @@ void kernel(float input[], float output[], float scale, float alpha)
     #pragma clang loop unroll_count(1) vectorize(disable)//vectorize_width(4)
     for (int i = 0; i < NTAPS; i++) {
         float x = input[i];
-        output[i] = scale * (max(0.0, x)) + min(0.0, alpha * (exp(x) - 1.0));
+        output[i] = scale * (max((float)(0.0), x)) + min((float)(0.0), alpha * (exp(x) - (float)(1.0)));
     }
 }
