+++ ops/activation/softplus.cpp
@@ -21,6 +21,6 @@ void kernel(float input[], float output[], float beta)
     #pragma clang loop unroll_count(1) vectorize(disable)//vectorize_width(4)
     for (int i = 0; i < NTAPS; i++) {
         float x = input[i];
-        output[i] = (1.0 / beta) * log(exp(beta * x));
+        output[i] = ((float)(1.0) / beta) * log(exp(beta * x));
     }
 }
