+++ ops/activation/softsign.cpp
@@ -20,6 +20,6 @@ void kernel(float input[], float output[])
     #pragma clang loop unroll_count(1) vectorize(disable)//vectorize_width(4)
     for (int i = 0; i < NTAPS; i++) {
         float x = input[i];
-        output[i] = x / (1.0 + abs(x));
+        output[i] = x / ((float)(1.0) + abs(x));
     }
 }
